refined the code to be more robuse and readable
wI should define an interface for ModbusPoll to inherit from and use that interface to set the raw data recieved from the serial line to the MobusPoll object. beacuse I don't want to make the (byte[]mData) buffer public but I want the ModbusMaster class to be able to access this field. (maybe there are better solution, I should think about it)